refactor: minor refactor DRY
