+++ src/cmipld/cli/config.py
@@ -1,45 +1,14 @@
 
 import json
+from pathlib import Path
 from rich.syntax import Syntax
 import typer
-from cmipld.core.service.settings import ServiceSettings, UniverseSettings, ProjectSettings
-from pathlib import Path
+from cmipld.core.service.settings import SETTINGS_FILE, ServiceSettings, load_settings
 from rich import print
 import toml
 
 app = typer.Typer()
 
-SETTINGS_FILE = Path("./src/cmipld/core/service/settings.toml")
-
-def load_settings() -> ServiceSettings:
-    """Load the settings from the TOML file."""
-    if SETTINGS_FILE.exists():
-        return ServiceSettings.load_from_file(str(SETTINGS_FILE))
-    else:
-        typer.echo("Settings file not found. Creating a new one with defaults.")
-        default_settings = ServiceSettings(
-        universe=UniverseSettings(
-            github_repo="https://github.com/example/universe",
-            branch="main",
-            local_path="/local/universe",
-            db_name="universe.db"
-        ),
-        Project1=ProjectSettings(
-                project_name="Project1",
-                github_repo="https://github.com/example/project1",
-                branch="main",
-                local_path="/local/project1",
-                db_name="project1.db"
-            )
-        
-    )
-
-        default_settings.save_to_file(str(SETTINGS_FILE))
-        return default_settings
-
-def save_settings(settings: ServiceSettings):
-    """Save the settings to the TOML file."""
-    settings.save_to_file(str(SETTINGS_FILE))
 
 def get_nested_value(settings_dict: dict, key_path: str):
     """Navigate through nested dictionary keys using dot-separated key paths."""
@@ -100,7 +69,8 @@ def config(key: str |None = typer.Argument(None), value: str|None = typer.Argume
         if selected_value:
             new_settings_dict = set_nested_value(json.loads(settings.model_dump_json()),key, value )
             new_settings = ServiceSettings(**new_settings_dict)
-            save_settings(new_settings)
+            new_settings.save_to_file(str(SETTINGS_FILE)) #TODO improved that .. remove SETTINGS_FILE dependancy
+            # save_settings(new_settings)
             typer.echo(f"New settings {new_settings.model_dump_json(indent=4)}")
             typer.echo(f"Updated '{key}' to '{value}'.")
         else:
